stying

+++ client/src/components/CommentCard.vue
@@ -12,8 +12,12 @@ const account = computed(() => AppState.account)
 
 async function deleteComment(commentId) {
   try {
+    const wantsToDestroy = await Pop.confirm('Are you sure you want to delete your comment?')
+    if (!wantsToDestroy) return
+
     await commentsService.deleteComment(commentId)
-  } catch (error) {
+  }
+  catch (error) {
     Pop.toast('Could not delete comments', 'error')
     console.error(error)
   }
@@ -24,32 +28,34 @@ async function deleteComment(commentId) {
 
 
 <template>
-  <div class="card mb-3">
-    <section class="container">
+  <div class="row mt-5">
+    <div class="card mb-3">
+      <section class="container">
 
 
-      <div class="row justify-content-start align-items-center">
-        <div class="col-2">
-          <img :src="comment.creator.picture" class="comment-image" alt="">
+        <div class="row justify-content-start align-items-center">
+          <div class="col-3">
+            <img :src="comment.creator.picture" class="comment-image" alt="">
+          </div>
+          <div class="col-8  py-3">
+            <p class="fw-medium text-start mb-0">
+              {{ comment.creator.name }}
+            </p>
+            <p class="text-start mb-0">
+              {{ comment.body }}
+            </p>
+          </div>
         </div>
-        <div class="col-8 py-3">
-          <p class="fw-medium text-start mb-0">
-            {{ comment.creator.name }}
-          </p>
-          <p class="text-start mb-0">
-            {{ comment.body }}
-          </p>
+        <div v-if="comment.creatorId == account?.id">
+          <div class="row text-end align-items-center">
+            <small><i @click="deleteComment(comment.id)" class="mdi mdi-delete text-danger opacity-50"></i></small>
+          </div>
         </div>
-      </div>
-      <div v-if="comment.creatorId == account?.id">
-        <div class="row text-end align-items-center">
-          <small><i @click="deleteComment(comment.id)" class="mdi mdi-delete text-danger opacity-50"></i></small>
-        </div>
-      </div>
 
 
 
-    </section>
+      </section>
+    </div>
   </div>
 </template>
