+++ client/src/pages/AccountPage.vue
@@ -25,8 +25,12 @@ async function getMyEvents() {
 
 async function deleteTicket(ticketId) {
   try {
+    const wantsToDestroy = await Pop.confirm('Are you sure you want to delete your ticket to this exclusive event?')
+    if (!wantsToDestroy) return
+
     await ticketsService.deleteTicket(ticketId)
-  } catch (error) {
+  }
+  catch (error) {
     Pop.toast('Could not delete your ticket', 'error')
     console.error(error)
   }
@@ -47,29 +51,36 @@ async function deleteTicket(ticketId) {
 </script>
 
 <template>
-  <div class="about text-center">
-    <div v-if="account">
-      <h1>Welcome {{ account.name }}</h1>
-      <img class="rounded" :src="account.picture" alt="" />
-      <p>{{ account.email }}</p>
+  <!-- I'd be interested in learning how to make the content here align with the content in the container below, which is not fluid. I want the bg color here to go to the edges though -->
+  <div class="container-fluid bg-secondary">
+    <div class="row my-4 px-5 py-3">
+      <div v-if="account" class="d-flex align-content-center">
+        <img class="me-3 rounded" :src="account.picture" alt="" />
+        <div class="my-auto">
+          <h6 class="mb-0">Welcome</h6>
+          <h1 class="">{{ account.name }}</h1>
 
-      <div class="container">
-        <div class="row">
-          <section v-for="ticket in accountTickets" :key="ticket.id" class="col-6 col-md-4">
-            <!-- FIXME reference Mick's album collaborations on his account page for delete -->
-            <TowerEventCard :towerEvent="ticket.event" />
-
-            <button @click="deleteTicket(ticket.id)" class="btn btn-outline-danger">Delete Ticket</button>
-
-            <!-- {{ accountTickets. }} -->
-          </section>
         </div>
       </div>
+      <div v-else>
+        <h1>Loading... <i class="mdi mdi-loading mdi-spin"></i></h1>
+      </div>
     </div>
+  </div>
+  <div>
+    <div class="container">
+      <div class="row">
+        <section v-for="ticket in accountTickets" :key="ticket.id" class="col-6 col-md-4">
+          <!-- FIXME reference Mick's album collaborations on his account page for delete -->
+          <TowerEventCard :towerEvent="ticket.event" />
+
+          <button @click="deleteTicket(ticket.id)" class="btn btn-outline-danger">Delete Ticket</button>
+
+          <!-- {{ accountTickets. }} -->
+        </section>
+      </div>
 
 
-    <div v-else>
-      <h1>Loading... <i class="mdi mdi-loading mdi-spin"></i></h1>
     </div>
   </div>
 </template>
